Move types to upper level dir
