+++ src/features/entity-download/file-handlers/experimental-bouton-density.ts
@@ -2,10 +2,9 @@
 import { getExperimentalBoutonDensity } from '@/api/entitycore/queries';
 import { EntityTypeEnum } from '@/api/entitycore/types';
 import { Metadata } from '@/features/entity-download/metadata';
+import { ExperimentalBoutonDensityJsonMetadata } from '@/features/entity-download/types';
 import { createTemplateFileEntry, getMetadataCsvEntryBase } from '@/features/entity-download/utils';
 import { WorkspaceContext } from '@/types/common';
-
-import { ExperimentalBoutonDensityJsonMetadata } from './types';
 
 export async function* getExperimentalBoutonDensityFiles(
   entityIds: string[],
