+++ src/features/entity-download/file-handlers/experimental-synapses-per-connection.ts
@@ -2,10 +2,9 @@
 import { getExperimentalSynapsesPerConnection } from '@/api/entitycore/queries';
 import { EntityTypeEnum } from '@/api/entitycore/types';
 import { Metadata } from '@/features/entity-download/metadata';
+import { ExperimentalSynapsesPerConnectionJsonMetadata } from '@/features/entity-download/types';
 import { createTemplateFileEntry, getMetadataCsvEntryBase } from '@/features/entity-download/utils';
 import { WorkspaceContext } from '@/types/common';
-
-import { ExperimentalSynapsesPerConnectionJsonMetadata } from './types';
 
 export async function* getExperimentalSynapsesPerConnectionFiles(
   entityIds: string[],
