+++ src/features/entity-download/file-handlers/memodel.ts
@@ -3,14 +3,13 @@ import { getEModel, getMEModel, getReconstructionMorphology } from '@/api/entity
 import { EntityTypeEnum } from '@/api/entitycore/types';
 import { ASSET_BASE_PATH } from '@/features/entity-download/constants';
 import { Metadata } from '@/features/entity-download/metadata';
+import { MemodelJsonMetadata } from '@/features/entity-download/types';
 import {
   createAssetFileEntry,
   createTemplateFileEntry,
   getMetadataCsvEntryBase,
 } from '@/features/entity-download/utils';
 import { WorkspaceContext } from '@/types/common';
-
-import { MemodelJsonMetadata } from './types';
 
 export async function* getMEmodelFiles(entityIds: string[], ctx?: WorkspaceContext) {
   const metadata = new Metadata<MemodelJsonMetadata>();
