+++ src/features/entity-download/file-handlers/reconstruction-morphology.ts
@@ -3,14 +3,13 @@ import { getReconstructionMorphology } from '@/api/entitycore/queries';
 import { EntityTypeEnum } from '@/api/entitycore/types';
 import { ASSET_BASE_PATH } from '@/features/entity-download/constants';
 import { Metadata } from '@/features/entity-download/metadata';
+import { ReconstructionMorphologyJsonMetadata } from '@/features/entity-download/types';
 import {
   createAssetFileEntry,
   createTemplateFileEntry,
   getMetadataCsvEntryBase,
 } from '@/features/entity-download/utils';
 import { WorkspaceContext } from '@/types/common';
-
-import { ReconstructionMorphologyJsonMetadata } from './types';
 
 export async function* getReconstructionMorphologyFiles(
   entityIds: string[],
