+++ src/features/entity-download/file-handlers/single-neuron-synaptome.ts
@@ -4,14 +4,13 @@ import { getSingleNeuronSynaptome } from '@/api/entitycore/queries/model/single-
 import { EntityTypeEnum } from '@/api/entitycore/types';
 import { ASSET_BASE_PATH } from '@/features/entity-download/constants';
 import { Metadata } from '@/features/entity-download/metadata';
+import { SingleNeuronSynaptomeJsonMetadata } from '@/features/entity-download/types';
 import {
   createAssetFileEntry,
   createTemplateFileEntry,
   getMetadataCsvEntryBase,
 } from '@/features/entity-download/utils';
 import { WorkspaceContext } from '@/types/common';
-
-import { SingleNeuronSynaptomeJsonMetadata } from './types';
 
 export async function* getSingleNeuronSynaptomeFiles(entityIds: string[], ctx?: WorkspaceContext) {
   const metadata = new Metadata<SingleNeuronSynaptomeJsonMetadata>();
