--- src/features/entity-download/file-handlers/types.ts
@@ -1,31 +1,0 @@
-export type ElectricalCellRecordingJsonMetadata = {
-  [key: string]: any;
-};
-
-export type EmodelJsonMetadata = {
-  [key: string]: any;
-};
-
-export type ExperimentalBoutonDensityJsonMetadata = {
-  [key: string]: any;
-};
-
-export type ExperimentalNeuronDensityJsonMetadata = {
-  [key: string]: any;
-};
-
-export type ExperimentalSynapsesPerConnectionJsonMetadata = {
-  [key: string]: any;
-};
-
-export type MemodelJsonMetadata = {
-  [key: string]: any;
-};
-
-export type ReconstructionMorphologyJsonMetadata = {
-  [key: string]: any;
-};
-
-export type SingleNeuronSynaptomeJsonMetadata = {
-  [key: string]: any;
-};
