+++ src/features/entity-download/types.ts
@@ -19,3 +19,35 @@ export type CsvEntry = CsvEntryBase & {
   idx: number;
   data_path?: string;
 };
+
+export type ElectricalCellRecordingJsonMetadata = {
+  [key: string]: any;
+};
+
+export type EmodelJsonMetadata = {
+  [key: string]: any;
+};
+
+export type ExperimentalBoutonDensityJsonMetadata = {
+  [key: string]: any;
+};
+
+export type ExperimentalNeuronDensityJsonMetadata = {
+  [key: string]: any;
+};
+
+export type ExperimentalSynapsesPerConnectionJsonMetadata = {
+  [key: string]: any;
+};
+
+export type MemodelJsonMetadata = {
+  [key: string]: any;
+};
+
+export type ReconstructionMorphologyJsonMetadata = {
+  [key: string]: any;
+};
+
+export type SingleNeuronSynaptomeJsonMetadata = {
+  [key: string]: any;
+};
